Separate vector layers, styles, and expression classes
