+++ gbdxtools/vector_styles.py
@@ -1,148 +1,4 @@
-import uuid
-import json
-
-class VectorLayer(object):
-    """ Represents a vector layer created from a geojson source, and knows how
-    to render itself as javascript.
-    """
-
-    def __init__(self, styles=None, **kwargs):
-        """ Abstract constructor for vector layers
-
-        Args:
-            styles (list): list of styles for which to create layers
-        """
-        self.source_id = uuid.uuid4().hex
-        if styles is not None:
-            self.styles = styles
-        else:
-            # nothing defined, so give them defaults
-            self.styles = [CircleStyle(**kwargs), LineStyle(**kwargs), FillStyle(**kwargs)]
-
-    def _layer_def(self, style):
-        """
-        Constructs a layer def with the proper fields
-            - implemented in subclasses 
-        Returns:
-            layer (dict): a layer json dict used for adding to maps 
-        """
-        raise NotImplementedError()
-
-    def _datasource_def(self):
-        """
-        Constructs a datasource def appropriate for the layer type
-            - implemented in subclasses
-        Returns
-            datasource (dict): a datasource json dict used for adding data to maps
-        """
-        raise NotImplementedError()
-
-    @property
-    def datasource(self):
-        """
-        Renders the datasource to add to the map, referenced by the layers
-        created by this layer instance.
-
-        Returns:
-            datasource (dict): a datasource json dict used for adding data to maps
-        """
-        return {'id': self.source_id, 'data': self._datasource_def()}
-
-    @property
-    def layers(self):
-        """
-        Renders the list of layers to add to the map
-        Returns:
-            list of layer entries suitable for use in mapbox-gl 'map.addLayer()' call
-        """
-        layers = []
-        layers = [self._layer_def(style) for style in self.styles]
-        return layers
-
-
-class VectorFeatureLayer(VectorLayer):
-    """ Represents a vector layer created from a geojson source, and knows how
-    to render itself as javascript.
-    """
-    def __init__(self, geojson, **kwargs):
-        """ Create a new VectorLayer
-
-        Args:
-            geojson (dict): a list of geojson features to render
-            styles (list): A list of style objects to be applied to the layer
-
-        Returns:
-            An instance of the VectorLayer
-        """
-        super(VectorFeatureLayer, self).__init__(**kwargs)
-        self.geojson = geojson
-
-    def _datasource_def(self):
-        return {'type': 'geojson', 'data': self.geojson}
-
-    def _layer_def(self, style):
-        return {
-            'id': type(style).__name__, # TODO - make this unique in the various styles
-            'type': style.type,
-            'source': self.source_id,
-            'paint': style.paint()
-        }
-        
-
-
-class VectorTileLayer(VectorLayer):
-    """ Represents a vector layer in a tile map, and knows how to render
-    itself as javascript.
-    """
-    def __init__(self, url=None, source_layer_name='GBDX_Task_Output', **kwargs):
-        """ Create a new VectorLayer
-
-        Args:
-            url (str): a vector tile url template
-            source_name (str): the name of the source layer in the vector tiles
-            styles (list): A list of style objects to be applied to the layer
-
-        Returns:
-            An instance of the VectorLayer
-        """
-        super(VectorTileLayer, self).__init__(**kwargs)
-        self.url = url
-        self.source_layer_name = source_layer_name
-
-    def _datasource_def(self):
-        return {'type': 'vector', 'tiles': [self.url]}
-
-    def _layer_def(self, style):
-        return {
-            'id': type(style).__name__, # TODO - make this unique in the various styles
-            'type': style.type,
-            'source': self.source_id,
-            'source-layer':  self.source_layer_name,
-            'paint': style.paint()
-        }
-
-class ImageLayer(object):
-    """
-      A layer for rendering images and image arrays to slippy maps
-    """
-    def __new__(self, image, coordinates):
-        """ Create a new ImageLayer
-
-        Args:
-            image (str): a vector tile url template
-            coordinates: the coordinate bounds (list of polygon corners) for placing the image
-        Returns:
-            An string of the layer definition
-        """
-        return json.dumps({
-          "id": uuid.uuid4().hex,
-          "type": "raster",
-          "source": {
-              "type": "image", 
-              "url": image, 
-              "coordinates": coordinates
-          }
-        })
+from gbdxtools.vector_style_expressions import StyleExpression
 
 
 class VectorStyle(object):
@@ -308,126 +164,3 @@ class FillStyle(VectorStyle):
             snippet['fill-translate'] = self.translate
 
         return snippet
-
-
-class StyleExpression(object):
-    """
-    Represents a mapbox-gl conditional for styling, knowing how
-    to turn itself into the appropriate javascript for use in
-    styling a layer.
-    """
-
-    def _expression_def(self):
-        """
-        Creates the list values which will be passed as a mapbox-gl
-        conditional expression.
-
-        Returns:
-            list of values for mapbox-gl conditional
-        """
-        raise NotImplementedError()
-
-    @property
-    def expression(self):
-        return self._expression_def()
-
-
-class MatchExpression(StyleExpression):
-    """
-    Represents a mapbox-gl "match" conditional expression, where a set of values
-    is matched against a property and styling applied based on the match.
-    """
-
-    def __init__(self, property_name=None, values=None, default_value=None):
-        """
-        Creates a match conditional expression, matching against the supplied property
-        name in a feature.  Values to match against and the styling to apply for those
-        matches are supplied in the 'values' dict.  If the feature property does not
-        match any of the values provided, it will be styled with the provided default.
-
-        Parameters:
-            property_name (str): the name of the feature property to match values against
-            values (dict): key/value pairs for property values and the style value to apply
-            default_value (?): default value to apply when no value matches
-        """
-        self.property_name = property_name
-        self.values = values
-        self.default_value = default_value
-
-    def _expression_def(self):
-        cond = ['match', ['get', self.property_name]]
-        for key in self.values:
-            cond.append(key)
-            cond.append(self.values[key])
-        cond.append(self.default_value)
-        return cond
-
-
-class InterpolateExpression(StyleExpression):
-    """
-    Represents a mapbox-gl "interpolate" expression, creating a smooth set of results
-    between provided stops for a given feature property.
-    """
-
-    def __init__(self, property_name=None, stops=None, type=None):
-        """
-        Creates an interpolate expression, based on values in the supplied property
-        of a feature.  Values indicating range boundaries and the styling to apply to features
-        in those ranges are supplied in the 'stops' dict.
-
-        The 'type' params must be a list whose values depend on the interpolation type
-        (taken from the mapbox-gl documentation):
-            - ["linear"]: interpolates linearly between the pair of stops just less than
-                          and just greater than the input.
-            - ["exponential", base]: interpolates exponentially between the stops just less
-                                     than and just greater than the input. base controls the
-                                     rate at which the output increases: higher values make
-                                     the output increase more towards the high end of the range. With values close to 1 the output increases linearly.
-            - ["cubic-bezier", x1, y1, x2, y2]: interpolates using the cubic bezier curve
-                                                defined by the given control points.
-
-        Parameters:
-            property_name (str): the name of the feature property to match values against
-            stops (dict): key/value pairs for range values and the style value to apply to values in that range
-            type (list): interpolation type params
-        """
-        self.property_name = property_name
-        self.stops = stops
-        self.type = type
-
-    def _expression_def(self):
-        cond = ['interpolate', self.type, ['get', self.property_name]]
-        for key in sorted(self.stops):
-            cond.append(key)
-            cond.append(self.stops[key])
-        return cond
-
-
-class StepExpression(StyleExpression):
-    """
-    Represents a mapbox-gl "step" expression, creating discrete stepped results
-    between provided stops for a given feature property.
-    """
-
-    def __init__(self, property_name=None, stops=None):
-        """
-        Creates a step expression, based on values in the supplied property
-        of a feature.  Values indicating range boundaries and the styling to apply to features
-        in those ranges are supplied in the 'stops' dict.
-
-
-        Parameters:
-            property_name (str): the name of the feature property to match values against
-            stops (dict): key/value pairs for range values and the style value to apply to values in that range
-        """
-        self.property_name = property_name
-        self.stops = stops
-
-    def _expression_def(self):
-        cond = ['step', ['get', self.property_name]]
-        keys = list(sorted(self.stops))
-        cond.append(self.stops[keys[0]])  # first entry must be the starting style value
-        for key in keys[1:]:
-            cond.append(key)
-            cond.append(self.stops[key])
-        return cond
